refactor(webapp): move giftcard content to a dedicated bounded-context
