+++ src/main/webapp/app/giftcard/domain/GiftCard.ts
