+++ src/main/webapp/app/giftcard/domain/GiftCardRepository.ts
@@ -1,4 +1,4 @@
-import type { GiftCard } from '@/home/domain/giftcard/GiftCard.ts';
+import type { GiftCard } from './GiftCard.ts';
 
 export interface GiftCardRepository {
   findAll(): Promise<GiftCard[]>;
