+++ src/main/webapp/app/home/infrastructure/primary/HomepageVue.vue
@@ -45,7 +45,7 @@
 </template>
 
 <script lang="ts">
-import GiftCardList from '@/home/domain/giftcard/GiftCardList.vue';
+import GiftCardList from '@/giftcard/infrastructure/primary/GiftCardList.vue';
 import { defineComponent } from 'vue';
 
 export default defineComponent({
