+++ src/main/webapp/app/shared/barcode/infrastructure/primary/VisualBarcode.vue
